:ghost: consolidated REST client.
